refactor(web): group cms components (part 3) - #95
