--- apps/web/src/lib/components/Link/LinkButton.svelte
@@ -1,14 +1,0 @@
-<script lang="ts">
-    import type { LinkOptions } from './types';
-
-    export let options: LinkOptions;
-</script>
-
-<div
-    role="button"
-    class="button"
->
-    <slot name="icon" />
-
-    {options.text ?? ''}
-</div>
